pyproject.toml
[tool.poetry]
name = "snitchbot"
version = "0.1.0"
description = ""
authors = ["Stepan Anokhin <stepan.anokhin@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
python-telegram-bot = "^21.1.1"
aiohttp = "^3.9.5"
aiodns = "^3.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
